1. Improved ptfce command line executable
2. Local smoothness estimation added
	a) SPM RPV
	b) FSL voxel based smoothness estimation

3. New testdata provided

4. Added simulated data generator

+++ grfClust.h
@@ -4,7 +4,7 @@
 
 double dvox(double h);
 double pvox(double h);
-double Es(double h, unsigned int V, double Rd);
+double Es(double h, double V, double Rd);
 double dcl(double h, void * p);
 double dclust(double h, void * p);
 double dvox_dclust(double h, void * p);
@@ -13,7 +13,7 @@ double pvox_clust(double actH, void * p);
 
 
 struct dcl_params{
-    unsigned long V;
+    double V;
     double Rd;
     double c;
     double ZestThr;
